Refactor: Move forensics modules to forensics directory
