+++ src/rldk/forensics/log_scan.py
